nfs: try to re-use transfer class on client retry
If door failed to receive redirect with in allowed time window it will kill
mover and forget about transfer. In case of WRITE a new file will be created
and location will be updated in the namespace. The following attempt from
the client to get a pool will create a read mover and all WRITE ops will
fail with ERR_PERM.

To avoid such situations try to re-use existing transfer class and bind cleanup
procedure to state disposal, which happens on close or when nfs server
cleans dead clients.

Fixes: https://github.com/dCache/dcache/issues/1846

Acked-by:
Target: master
Require-book: no
Require-notes: no