chimera: introduce truncate method
Motivation:
In some workloads end-users need to overwrite existing files. As this is
not possible with dcache, most of the doors have implement it as combination
of remove+create.

Since we have inumber as yet another level of indirection, overwriting
a file can be implemented as truncate.

The truncate itself is implemented as 'killing' the old pnfsid and assigning
a new one with the same inumber. This will allow NFS clients still be
happy, as nfs file handle is constructed from inumber.

Modification:
Add truncate method which will simulate remove to clean location and levels.

Result:
no vidible changes yet

Acked-by:
Target: master
Require-book: no
Require-notes: no