chimera: introduce truncate method
Motivation:
In some workloads end-users need to overwrite existing files. As this is
not possible with dcache, most of the doors simulate overwrite as combination
of remove+create.

Since we have inumber as yet another level of indirection, overwriting
of a file can be implemented as truncate.

The truncate itself is implemented as 'killing' the old pnfsid and assigning
a new pnfsid to the existing inumber. This will allow NFS clients still be
happy, as nfs file handle is constructed from inumber.

Modification:
Add truncate method which will simulate remove to clean location and levels.

Result:
no visible changes yet

NOTICE: as truncate method changes inumber<=>pnfsid mapping, the mapping
cache have to be invalidated. In deployments with multiple PnfsManagers
and NFS doors the under laying hazelcast cache have to be configured to
run in cluster mode.

Acked-by:
Target: master
Require-book: no
Require-notes: no