httpd: fix three small null-related bugs in accessing active transfer fields
Motivation:

Patch 9258 introduced a UserInfo object and also somewhat restructured TransferObserverV1.
In so doing it introduced two minor bugs.  The first was an erroneous suppression
of a condition preventing the addition of nulls to an argument list in the creation
of the ascii context info.  The latter has to do with the handling of exceptions thrown by
the Subjects static methods (this was observed to occur using anonymous dccp).

Also, a method called by the webadmin version of this service also neglected to
check for a null value.

Modification:

1.  Restore the condition check (mover == null) for the ascii context info.
2.  Provide for the handling of NoSuchElement and IllegalArgument exceptions
3.  Handle the null value for bytesTransferred.

Result:

Errors are now handled gracefully rather than causing Wicket to crash or a log
full of stack traces.

Target: master
Require-notes: no
Require-book: no
Acked-by: Gerd
(cherry picked from commit c39133d664b0ccedcdf975a326ff5383f90dfee6)

Conflicts:
	modules/dcache/src/main/java/diskCacheV111/cells/TransferObserverV1.java