Motivaton:
Pool selection may be based on directory storage tags. A move in chimera namespace
is purely rename operation. Therefore, a possibility of conflict exists
when files are "moved" from a source with tags that are used to direct data to
non-tape back-end pools to a destination directory with tags that direct
data to tape-backed pools. No data will be actually moved to tape. Moreover,
if a directory or the whole directory tree is moved, any new files written
to this directory or any directory underneath would not go to tape.

Modification:

Added variable

pnfsmanager.enable.move-to-directory-with-different-storageclass

true is default. If set to false, the source and destionation
storage class will be checked. If there is no match, the mv would fail.

Result:

Unable to move entries between directories having different storage class tags
`