namespace/chimera: Motivation:
A pool selection may be based on directory tags that form storage class and cache class. A move in the namespace is a purely rename operation. Therefore, a possibility of conflict exists when files/directories are moved from a source with tags that are used to direct data to non HSM-backed pools to a destination directory with tags that direct data to HSM-backed pools. No data will be actually moved to tape. Moreover, if a directory or the whole directory tree is moved, any new files written to this directory or any directory underneath would not go to tape. This behavior is deemed confusing by some of our customers.

Modification:

Introduced variables:

pnfsmanager.enable.move-to-directory-with-different-storageclass

with default value of true, and

nfs.enable.pnfsmanager-query-on-move

with default value of false.

Setting these variables to false and true correspondingly disables "mv" between source and destination having different storage and cache classes. When using NFS door, nfs.enable.pnfsmanager-query-on-move=true enables sending rename message to PnfsManager and pnfsmanager.enable.move-to-directory-with-different-storageclass=false triggers check for source and destination storage and cache classes and results in refusal to perform an "mv" if source and destination storage and cache classes do not match.

Result:

Cannot move files between directories having different storage/cache classes when nfs.enable.pnfsmanager-query-on-move is set to true and pnfsmanager.enable.move-to-directory-with-different-storageclass is set to false.

Release Notes:

A pool selection may be based on directory tags that form storage class and cache class. A move in the namespace is a purely rename operation. Therefore, a possibility of conflict exists when files/directories are moved from a source with tags that are used to direct data to non HSM-backed pools to a destination directory with tags that direct data to HSM-backed pools. No data will be actually moved to tape. Moreover, if a directory or the whole directory tree is moved, any new files written to this directory or any directory underneath would not go to tape. This behavior is deemed confusing by some of our customers.

We introduced two variables:

pnfsmanager.enable.move-to-directory-with-different-storageclass

with default value of true, and

nfs.enable.pnfsmanager-query-on-move

with default value of false.

Setting these variables to false and true correspondingly disables "mv" between source and destination having different storage and cache classes. When using NFS door, nfs.enable.pnfsmanager-query-on-move=true enables sending rename message to PnfsManager and pnfsmanager.enable.move-to-directory-with-different-storageclass=false triggers check for source and destination storage and cache classes and results in refusal to perform an "mv" if source and destination storage and cache classes do not match.

Having additional message sent from NFS door to PnfsManager adds latency to mv command and therefore is disabled by default. Setting this variabe to true without simultaneously setting pnfsmanager.enable.move-to-directory-with-different-storageclass to false is not practical as this is currently the only functionality it is used for.

Acked-by: Albert Rossi arossi@fnal.gov
Acked-by: Tigran Mkrtchyan tigran.mkrtchyan@desy.de
Target: master
Require-book: yes
Require-notes: yes
Patch: https://rb.dcache.org/r/9494/

merged what is equivalent of
 https://rb.dcache.org/r/9494/