alarms: only send email on first alarm occurrence; reset count history on reopened alarm
combines patches #9633 and #9647

Motivation:

9555, which modified the log entry handling to be able to use listener

extensions to process alarms, also modified the handling of alarms such
that this post-processing, as well as email, occurs only on the first
instance of potentially duplicate alarms.   This behavior is currently
not enforced for the stable branches with respect to alarm emails.

Modification:

Backport the change which affects the handling of alarm emails.

Result:

Alarm email notifications are sent only on the first occurrence
of given alarm (i.e., for that alarm instance's unique ID).

-------------------------

Motivation:

9555 (master) and #9633 (2.13-2.16)

corrected the behavior of the log entry handler such that emails
(and eventually plugin calls) are triggered only on the first
occurrence of an alarm (id).

This patch finishes that correction as regards alarms which are
in the closed state.  If an alarm has been closed and not deleted,
but then occurs again, the count history (the 'received' attribute)
should be reset to 1, in order to treat this as a new (set of)
occurrences, and to guarantee a new notification will be sent.

Modification:

Reset the received value to 1 if the alarm is being reopened.

Result:

Notification and triggering of plugins will occur on reopened
alarms.

Acked-by:  Dmitry
Acked-by:  Paul