poolmanager: re-implement request container
WIP: not all functionality is covered yet.

Motivation:
Request container is the hard of the pool selection. It responsible for
pool selection and, if required, stage and p2p, when data is not
available from the desired pool.

By being a very critical component, is lacks of test as well as code
base is not simple to understand and maintain.

Modification:
Implement a new drop-in replacement for request container. It has three
request queues:

  - fast track, for any action which can be done without blocking.
  - working queue, for requests which can't be served right away due
    to not available pools or missing replicas
  - waiting queue, for all requests that needan external event, like
    p2p, stage or operator intervention.

The fast track is implemented and 'same thread executor', which utilizes
the message thread.

Add unit test to cover typical state transitions.

Result:
simple and faster code.

Acked-by:
Target: master
Require-book: no
Require-notes: yes