telemetry: new cell to collect usage data
Motivation:

At the moment, information about dCache-instances are stored in BDII. As a new solution,
this cell should collect the information and store them in a centralized database.

Modification:

InstanceData contains attributes and methods to collect, hold and update different information
about the instance. SendData updates and sends the data once in an hour to a central database.
Therefore it uses a scheduledExecutor. As database, elasticsearch is used with a new index per day.

Result:

Information about dCache-instances are collected and stored in a database if the cell is enabled.
It's possible to add things like contact information to the collected data in the future.

Acked-by:
Target: master
Require-book: yes
Require-notes: yes
Patch: https://rb.dcache.org/r/12202/